Fix unsigned difference expression compared to zero
Fixes: https://github.com/OpenSCAP/openscap/security/code-scanning/1414

+++ src/XCCDF_POLICY/xccdf_policy_remediate.c
@@ -790,7 +790,7 @@ static inline int _parse_ansible_fix(const char *fix_text, struct oscap_list *va
 	int ovector[9];
 
 	const size_t fix_text_len = strlen(fix_text);
-	int start_offset = 0;
+	size_t start_offset = 0;
 	while (true) {
 		const int match = oscap_pcre_exec(re, fix_text, fix_text_len, start_offset,
 				0, ovector, sizeof(ovector) / sizeof(ovector[0]));
